failed attempt to minimize Box<T> in generated types
Deciding what types to box is directly reducible to minimum feedback arc
set, which is NP hard. Initially I misjudged how many total types there
were in the bot api and attempted an exact solution using dynamic
programming (tested in this commit). Time complexity is O(n2^n) so the
runtime is prohibitively long.

My plan is to retry with an approximation